+++ dominatorarena-api/src/main/java/com/biel/dominatorarena/api/responses/WorkBlockResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
  * Created by Biel on 4/12/2016.
  */
 public class WorkBlockResponse {
-    int workBlockId;
+    Long workBlockId;
     //Files
     List<ConfigurationResponse> configurationResponses;
     List<StrategyVersionResponse> strategyVersionResponses;
@@ -16,18 +16,18 @@ public class WorkBlockResponse {
     protected WorkBlockResponse() {
     }
 
-    public WorkBlockResponse(int workBlockId, List<ConfigurationResponse> configurationResponses, List<StrategyVersionResponse> strategyVersionResponses, List<BattleResponse> battleResponses) {
+    public WorkBlockResponse(Long workBlockId, List<ConfigurationResponse> configurationResponses, List<StrategyVersionResponse> strategyVersionResponses, List<BattleResponse> battleResponses) {
         this.workBlockId = workBlockId;
         this.configurationResponses = configurationResponses;
         this.strategyVersionResponses = strategyVersionResponses;
         this.battleResponses = battleResponses;
     }
 
-    public int getWorkBlockId() {
+    public Long getWorkBlockId() {
         return workBlockId;
     }
 
-    public void setWorkBlockId(int workBlockId) {
+    public void setWorkBlockId(Long workBlockId) {
         this.workBlockId = workBlockId;
     }
